1-list_len.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int i;
+	size_t i;
 
 	if (h == NULL)
 		return (0);
@@ -17,9 +17,9 @@ size_t print_list(const list_t *h)
 		if (h->str == NULL)
 			printf("[0] (nil)\n");
 		else
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%lu] %s\n", h->len, h->str);
 		h = h->next;
 	}
-	printf("[%d] %s\n", h->len, h->str);
+	printf("[%lu] %s\n", h->len, h->str);
 	return (i);
 }
